requirements.txt
numpy==1.26.4
pandas==2.2.3
pillow==11.0.0
tqdm==4.66.6
einops==0.8.0
addict==2.4.0
yapf==0.40.2
matplotlib==3.9.2
pydantic==2.9.2
scikit-learn==1.5.2

open_clip_torch==2.29.0
bitsandbytes==0.44.1
accelerate==1.1.0
safetensors==0.4.5
transformers==4.48.0
diffusers==0.31.0
peft==0.13.2
datasets==3.1.0
timm==1.0.11
pycocotools==2.0.8
openai==1.59.3

git+https://github.com/facebookresearch/segment-anything.git